Use our own `dicthash` function in cacheing key_value and row transforms
It turns out that relying on the id of a dict - especially one that is
recreated at search time rather than being static - fails significantly.
So we implement our own 'dicthash' function that uses a CRC-32 on the
concatenated, sorted key-value pairs (encoded from UTF-8 to bytes) of
the dict.  For anything else, it can use the `id()` value of the object.

Signed-off-by: Paul Wayper <paulway@redhat.com>