Rework of keyword_search and key_match caching
After a bit of testing, we discovered that caching the result of
`keyword_search` does not speed up the results delivery.  However, a
cache on the transformed row does speed things up considerably.  And
after extensive testing, a cache on `id(rows) + id(row)` gives the best
combination of performance and security against accidental reuse of row
memory (which happens a lot in tests).

This comes with a specific caveat - that it's not impossible to reuse
both the row dict and the containing rows object/list.  Previously, the
`pam` parser module rewrote the row dicts each time, which confused the
cache a lot (and was also inefficient).  This is most likely to be
picked up in tests (because the parser tests revealed the failure points
in previous algorithms).

The other improvement is a fast-path for straight equality comparisons.

Signed-off-by: Paul Wayper <paulway@redhat.com>