Cache row transforms on parent object, if we can find it
After a lot of testing we've discovered that an internal cache dictionary
doesn't work.  Sometimes the 'rows' parameter is a list, which we can't add
attributes to; sometimes it's a list-like object with dictionaries which we
can't add attributes to; sometimes it's a list-like object with dict-like
objects.  We have to return the original row, rather than any transformed
row.

Testing has shown it's better to incur the up-front cost of transforming the
rows in the outer loop.  If we can attach that list of transformed rows to
an object - the `rows` object, or a parent object if `rows` is a list - then
that speeds up all future searches.  We add the optional `parent` parameter to
the `keyword_search()` function, so that parsers can supply that to attach
the transformed rows to.

Signed-off-by: Paul Wayper <paulway@redhat.com>