Cache the keyword transforms, pre-prepare the search terms
This introduces two further speed improvements.  The first is that instead of
constructing a new list of dictionaries for the transformed rows, we can
simply construct a dictionary mapping the search term keyword (e.g.
`command_name`) to the row data key (e.g. `command name`).  This is much
smaller, involves less of a full table scan, and is easy to substitute into
our search process.

The second is that we can actually pre-prepare the search terms into a list
of tuples containing the row data key, the matcher function name and function,
and the value sought.  Instead of doing this for every row in the data, we
do it ahead of time.

A corollary of both of these is that we can check if the search term does not
occur anywhere in the row data keys, before we even do a search of the data.
While this would be rare, it could occur with some data that has optional
fields (e.g. CPU info for different architectures).

Signed-off-by: Paul Wayper <paulway@redhat.com>